requirements.txt
setuptools
wheel
requests
locustio>=0.11.0
locust-fixed-interval
requests
PyHamcrest
pytest
behave
PyHamcrest
Faker
bzt
pyyaml